Use a single body data processor for testing
